Updates courses form and route to reflect current user logged in
Also updates sample data for admin user to include a course and a book

+++ app/courses/routes.py
@@ -1,5 +1,5 @@
 from flask import render_template, redirect, request, url_for
-from flask_login import login_required
+from flask_login import login_required, current_user
 
 from app import db
 from app.courses import bp
@@ -10,7 +10,7 @@ from .forms import CourseEditForm, CourseAddForm
 @bp.route('/index', methods = ['GET', 'POST'])
 @login_required
 def index():
-    courses = Course.query.all()
+    courses = Course.query.filter_by(user_id = current_user.id)
     return render_template (
         'courses_list.html',
         title='Courses',
@@ -21,6 +21,10 @@ def index():
 @login_required
 def edit(id):
     course = Course.query.get_or_404(id)
+    # Double check that the user trying to edit this course actually
+    # owns the course that is being edited (if not, this is unauthorised)
+    if course.user_id != current_user.id:
+        return render_template('unauthorized.html'), 401
     form = CourseEditForm(obj=course)
     page_title = f'Editing {course.title}'
     if form.validate_on_submit():
@@ -40,6 +44,9 @@ def edit(id):
 @login_required
 def view(id):
     course = Course.query.get_or_404(id)
+    # Double check that the user accessing this course actually owns it
+    if course.user_id != current_user.id:
+        return render_template('unauthorized.html'), 401
     return render_template('course_details.html', title = course.title, course = course)
 
 @bp.route('/add_course', methods = ['GET','POST'])
@@ -59,6 +66,9 @@ def add():
         course.next_steps = form.next_steps.data
         course.barriers = form.barriers.data
         course.notes = form.notes.data
+
+        # Set the user of the course to the logged in user
+        course.user_id = current_user.id
 
         # Add the tags we need
         selected_tag_ids = form.tags.data
@@ -81,6 +91,12 @@ def add():
 @login_required
 def delete(id):
     course = Course.query.get_or_404(id)
+
+    # Double check that the user is trying to delete actually
+    # owns the course (if not, this is unauthorised)
+    if course.user_id != current_user.id:
+        return render_template('unauthorized.html'), 401
+
     course_title = course.title
     page_title = f'Deleted {course_title}'
     db.session.delete(course)
@@ -91,8 +107,14 @@ def delete(id):
 @login_required
 def edit_tags(id):
     course = Course.query.get_or_404(id)
+
+    # Double check that the user trying to edit tags
+    # owns the course that is being edited (if not, this is unauthorised)
+    if course.user_id != current_user.id:
+        return render_template('unauthorized.html'), 401
+
     course_tag_ids = [tag.id for tag in course.tags]
-    tags = Tag.query.all()
+    tags = Tag.query.filter_by(user_id = current_user.id)
 
     if request.method == 'POST':
         selected_tag_ids = request.form.getlist('tag-checkboxes')
